+++ packages/app-file-manager/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.13.0](https://github.com/webiny/webiny-js/compare/v4.13.0-beta.0...v4.13.0) (2020-10-06)
+
+**Note:** Version bump only for package @webiny/app-file-manager
+
+
+
+
+
+# [4.13.0-beta.0](https://github.com/webiny/webiny-js/compare/v4.12.1...v4.13.0-beta.0) (2020-10-06)
+
+
+### Bug Fixes
+
+* **app-file-manager:** add missing `security-scopes-list` plugin ([a3ceabd](https://github.com/webiny/webiny-js/commit/a3ceabdb99e6e0a4c401bcb820ab5ab0b0fe7917))
+
+
+
+
+
 ## [4.12.1](https://github.com/webiny/webiny-js/compare/v4.12.1-beta.0...v4.12.1) (2020-09-17)
 
 **Note:** Version bump only for package @webiny/app-file-manager